requirements.txt
pytest
selenium
webdriver-manager
dicttoxml
allure-pytest
